Welp API Key header is now needed. Will prompt user

+++ itemHashDestiny.py
@@ -4,21 +4,28 @@ import urllib
 import os
 import sys
 import json
-print "\nWriting itemHash text file to local disk..."
-url = 'http://www.destinydb.com/sitemap.xml/items/1'
-xml = urllib.urlopen(url).read()
-xmldoc = minidom.parseString(xml)
-loc_values = xmldoc.getElementsByTagName('loc')
-f=open('itemHashCatalog.txt','w')
-for loc_val in loc_values:
-	item=(loc_val.firstChild.nodeValue)
-	if len((str(item).split('items/')))==1:
-		pass
-	else:
-		#print (str(item).split('items/'))[1][:(str(item).split('items/'))[1].replace('-', ' ' ,1).index(' ')]+' - '+(str(item).split('items/'))[1][(str(item).split('items/'))[1].replace('-', ' ' ,1).index(' ')+1:].replace("-"," ").title()
-		f.write((str(item).split('items/'))[1][:(str(item).split('items/'))[1].replace('-', ' ' ,1).index(' ')]+' - '+(str(item).split('items/'))[1][(str(item).split('items/'))[1].replace('-', ' ' ,1).index(' ')+1:].replace("-"," ").title()+'\n')
-f.close()
-print '\nComplete! File saved to:\n'+os.getcwd()+'/itemHashCatalog.txt\n'
+
+#prompt user for api key
+print "\nChecking local disk in current directory..."
+if (os.path.exists('itemHashCatalog.txt')==False):
+	print "Catalog needed."
+	print "\nWriting itemHash text file to local disk..."
+	url = 'http://www.destinydb.com/sitemap.xml/items/1'
+	xml = urllib.urlopen(url).read()
+	xmldoc = minidom.parseString(xml)
+	loc_values = xmldoc.getElementsByTagName('loc')
+	f=open('itemHashCatalog.txt','w')
+	for loc_val in loc_values:
+		item=(loc_val.firstChild.nodeValue)
+		if len((str(item).split('items/')))==1:
+			pass
+		else:
+			#print (str(item).split('items/'))[1][:(str(item).split('items/'))[1].replace('-', ' ' ,1).index(' ')]+' - '+(str(item).split('items/'))[1][(str(item).split('items/'))[1].replace('-', ' ' ,1).index(' ')+1:].replace("-"," ").title()
+			f.write((str(item).split('items/'))[1][:(str(item).split('items/'))[1].replace('-', ' ' ,1).index(' ')]+' - '+(str(item).split('items/'))[1][(str(item).split('items/'))[1].replace('-', ' ' ,1).index(' ')+1:].replace("-"," ").title()+'\n')
+	f.close()
+	print '\nComplete! File saved to:\n'+os.getcwd()+'/itemHashCatalog.txt\n'
+else:
+	print "\nHash Catalog already exists in current directory!\n"
 
 print "For what do you want me to pull info for?"
 item = raw_input('')
